requirements.txt
twitchio<3
requests
python-dotenv
websockets